requirements.txt
flask
flask_bootstrap
pandas
wtforms
biopython(Bio.SeqUtils.ProtParam)